feat: move old migrations to new format
