+++ src/module/chat/index.ts
@@ -1,0 +1,14 @@
+import { GurpsModule } from '@gurps-types/gurps-module.js'
+
+import { migrations } from './migrations/index.js'
+
+function init() {
+  console.log('GURPS | Initializing GURPS Chat Module')
+}
+
+/* ---------------------------------------- */
+
+export const ChatModule: GurpsModule = {
+  init,
+  migrations,
+}
